move EthAddressFromPubKey to chain/types/ethtypes.
