+++ chain/wallet/key/key.go
@@ -1,6 +1,7 @@
 package key
 
 import (
+	"github.com/filecoin-project/lotus/chain/types/ethtypes"
 	"golang.org/x/xerrors"
 
 	"github.com/filecoin-project/go-address"
@@ -9,7 +10,6 @@ import (
 
 	"github.com/filecoin-project/lotus/chain/types"
 	"github.com/filecoin-project/lotus/lib/sigs"
-	"github.com/filecoin-project/lotus/lib/sigs/delegated"
 )
 
 func GenerateKey(typ types.KeyType) (*Key, error) {
@@ -54,7 +54,7 @@ func NewKey(keyinfo types.KeyInfo) (*Key, error) {
 		}
 	case types.KTDelegated:
 		// Assume eth for now
-		ethAddr, err := delegated.EthAddressFromPubKey(k.PublicKey)
+		ethAddr, err := ethtypes.EthAddressFromPubKey(k.PublicKey)
 		if err != nil {
 			return nil, xerrors.Errorf("failed to calculate Eth address from public key: %w", err)
 		}
